+++ aiosmtpd/controller.py
@@ -7,11 +7,18 @@ import os
 import ssl
 import threading
 import time
+from abc import ABCMeta, abstractmethod
 from contextlib import ExitStack
+from pathlib import Path
 from socket import AF_INET6, SOCK_STREAM, create_connection, has_ipv6
 from socket import socket as makesock
 from socket import timeout as socket_timeout
-from typing import Any, Coroutine, Dict, Optional
+
+try:
+    from socket import AF_UNIX
+except ImportError:  # pragma: on-not-win32
+    AF_UNIX = None
+from typing import Any, Coroutine, Dict, Optional, Union
 from warnings import warn
 
 from public import public
@@ -82,9 +89,13 @@ class _FakeServer(asyncio.StreamReaderProtocol):
 
 
 @public
-class Controller:
+class BaseThreadedController(metaclass=ABCMeta):
+    """
+    `Documentation can be found here
+    <https://aiosmtpd.readthedocs.io/en/latest/controller.html>`_.
+    """
     server: Optional[AsyncServer] = None
-    server_coro: Coroutine = None
+    server_coro: Optional[Coroutine] = None
     smtpd = None
     _factory_invoked: Optional[threading.Event] = None
     _thread: Optional[threading.Thread] = None
@@ -94,37 +105,31 @@ class Controller:
         self,
         handler,
         loop=None,
-        hostname=None,
-        port=8025,
         *,
-        ready_timeout=1.0,
-        ssl_context: ssl.SSLContext = None,
+        ready_timeout: float = 1.0,
+        ssl_context: Optional[ssl.SSLContext] = None,
         # SMTP parameters
-        server_hostname: str = None,
-        server_kwargs: Dict[str, Any] = None,
+        server_hostname: Optional[str] = None,
         **SMTP_parameters,
     ):
-        """
-        `Documentation can be found here
-        <http://aiosmtpd.readthedocs.io/en/latest/aiosmtpd\
-/docs/controller.html#controller-api>`_.
-        """
         self.handler = handler
-        self.hostname = get_localhost() if hostname is None else hostname
-        self.port = port
-        self.ssl_context = ssl_context
-        self.loop = asyncio.new_event_loop() if loop is None else loop
+        if loop is None:
+            self.loop = asyncio.new_event_loop()
+        else:
+            self.loop = loop
         self.ready_timeout = float(
             os.getenv("AIOSMTPD_CONTROLLER_TIMEOUT", ready_timeout)
         )
-        if server_kwargs:
+        self.ssl_context = ssl_context
+        self.SMTP_kwargs: Dict[str, Any] = {}
+        if "server_kwargs" in SMTP_parameters:
             warn(
                 "server_kwargs will be removed in version 2.0. "
                 "Just specify the keyword arguments to forward to SMTP "
                 "as kwargs to this __init__ method.",
                 DeprecationWarning,
             )
-        self.SMTP_kwargs: Dict[str, Any] = server_kwargs or {}
+            self.SMTP_kwargs = SMTP_parameters.pop("server_kwargs")
         self.SMTP_kwargs.update(SMTP_parameters)
         if server_hostname:
             self.SMTP_kwargs["hostname"] = server_hostname
@@ -150,6 +155,14 @@ class Controller:
         finally:
             self._factory_invoked.set()
 
+    @abstractmethod
+    def _create_server(self) -> Coroutine:
+        raise NotImplementedError  # pragma: nocover
+
+    @abstractmethod
+    def _trigger_server(self):
+        raise NotImplementedError  # pragma: nocover
+
     def _run(self, ready_event):
         asyncio.set_event_loop(self.loop)
         try:
@@ -157,14 +170,8 @@ class Controller:
             # detect the types of the vars. Cannot use `assert isinstance`, because
             # Python 3.6 in asyncio debug mode has a bug wherein CoroWrapper is not
             # an instance of Coroutine
-            srv_coro: Coroutine = self.loop.create_server(
-                self._factory_invoker,
-                host=self.hostname,
-                port=self.port,
-                ssl=self.ssl_context,
-            )
-            self.server_coro = srv_coro
-            srv: AsyncServer = self.loop.run_until_complete(srv_coro)
+            self.server_coro = self._create_server()
+            srv: AsyncServer = self.loop.run_until_complete(self.server_coro)
             self.server = srv
         except Exception as error:  # pragma: on-wsl
             # Usually will enter this part only if create_server() cannot bind to the
@@ -182,24 +189,6 @@ class Controller:
         self.loop.run_until_complete(self.server.wait_closed())
         self.loop.close()
         self.server = None
-
-    def _testconn(self):
-        """
-        Opens a socket connection to the newly launched server, wrapping in an SSL
-        Context if necessary, and read some data from it to ensure that factory()
-        gets invoked.
-        """
-        # IMPORTANT: Windows does not need the next line; for some reasons,
-        # create_connection is happy with hostname="" on Windows, but screams murder
-        # in Linux.
-        # At this point, if self.hostname is Falsy, it most likely is "" (bind to all
-        # addresses). In such case, it should be safe to connect to localhost)
-        hostname = self.hostname or get_localhost()
-        with ExitStack() as stk:
-            s = stk.enter_context(create_connection((hostname, self.port), 1.0))
-            if self.ssl_context:
-                s = stk.enter_context(self.ssl_context.wrap_socket(s))
-            _ = s.recv(1024)
 
     def start(self):
         assert self._thread is None, "SMTP daemon already running"
@@ -225,10 +214,9 @@ class Controller:
         # factory() go undetected. To trigger factory() invocation we need to open
         # a connection to the server and 'exchange' some traffic.
         try:
-            self._testconn()
+            self._trigger_server()
         except socket_timeout:
             # We totally don't care of timeout experienced by _testconn,
-            # which _will_ happen if factory() experienced problems.
             pass
         except Exception:
             # Raise other exceptions though
@@ -251,9 +239,111 @@ class Controller:
         for task in _all_tasks(self.loop):
             task.cancel()
 
-    def stop(self):
-        assert self._thread is not None, "SMTP daemon not running"
+    def stop(self, no_assert=False):
+        assert no_assert or self._thread is not None, "SMTP daemon not running"
         self.loop.call_soon_threadsafe(self._stop)
-        self._thread.join()
-        self._thread = None
+        if self._thread is not None:
+            self._thread.join()
+            self._thread = None
         self._thread_exception = None
+        self._factory_invoked = None
+        self.server_coro = None
+        self.server = None
+        self.smtpd = None
+
+
+@public
+class Controller(BaseThreadedController):
+    """
+    `Documentation can be found here
+    <https://aiosmtpd.readthedocs.io/en/latest/controller.html>`_.
+    """
+    def __init__(
+        self,
+        handler,
+        hostname: Optional[str] = None,
+        port: int = 8025,
+        loop=None,
+        *,
+        ready_timeout: float = 1.0,
+        ssl_context: ssl.SSLContext = None,
+        # SMTP parameters
+        server_hostname: Optional[str] = None,
+        **SMTP_parameters,
+    ):
+        super().__init__(
+            handler,
+            loop,
+            ready_timeout=ready_timeout,
+            server_hostname=server_hostname,
+            **SMTP_parameters
+        )
+        self.hostname = get_localhost() if hostname is None else hostname
+        self.port = port
+        self.ssl_context = ssl_context
+
+    def _create_server(self) -> Coroutine:
+        return self.loop.create_server(
+            self._factory_invoker,
+            host=self.hostname,
+            port=self.port,
+            ssl=self.ssl_context,
+        )
+
+    def _trigger_server(self):
+        """
+        Opens a socket connection to the newly launched server, wrapping in an SSL
+        Context if necessary, and read some data from it to ensure that factory()
+        gets invoked.
+        """
+        # At this point, if self.hostname is Falsy, it most likely is "" (bind to all
+        # addresses). In such case, it should be safe to connect to localhost)
+        hostname = self.hostname or get_localhost()
+        with ExitStack() as stk:
+            s = stk.enter_context(create_connection((hostname, self.port), 1.0))
+            if self.ssl_context:
+                s = stk.enter_context(self.ssl_context.wrap_socket(s))
+            _ = s.recv(1024)
+
+
+class UnixSocketController(BaseThreadedController):  # pragma: on-win32 on-cygwin
+    """
+    `Documentation can be found here
+    <https://aiosmtpd.readthedocs.io/en/latest/controller.html>`_.
+    """
+    def __init__(
+        self,
+        handler,
+        unix_socket: Optional[Union[str, Path]],
+        loop=None,
+        *,
+        ready_timeout=1.0,
+        ssl_context=None,
+        # SMTP parameters
+        server_hostname: str = None,
+        **SMTP_parameters,
+    ):
+        super().__init__(
+            handler,
+            loop,
+            ready_timeout=ready_timeout,
+            ssl_context=ssl_context,
+            server_hostname=server_hostname,
+            **SMTP_parameters
+        )
+        self.unix_socket = str(unix_socket)
+
+    def _create_server(self) -> Coroutine:
+        return self.loop.create_unix_server(
+            self._factory_invoker,
+            path=self.unix_socket,
+            ssl=self.ssl_context,
+        )
+
+    def _trigger_server(self):
+        with ExitStack() as stk:
+            s: makesock = stk.enter_context(makesock(AF_UNIX, SOCK_STREAM))
+            s.connect(self.unix_socket)
+            if self.ssl_context:
+                s = stk.enter_context(self.ssl_context.wrap_socket(s))
+            _ = s.recv(1024)